requirements.txt
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.2.0
stripe==6.4.0
sorl-thumbnail==12.9.0
openai
django-widget-tweaks==1.4.12
python-dotenv==1.0.0
